Refactor workflow code into files and folders
