+++ keepercommander/commands/workflow/__init__.py
@@ -5,23 +5,11 @@
 #              |_|
 #
 # Keeper Commander
-# Copyright 2024 Keeper Security Inc.
+# Copyright 2026 Keeper Security Inc.
 # Contact: ops@keepersecurity.com
 #
 
-"""
-Keeper PAM Workflow Commands
-
-This module implements commands for managing PAM workflows including:
-- Configuration management (create, update, delete workflows)
-- Approver management (add, remove approvers)
-- Workflow state inspection (get status, list requests)
-- Workflow actions (request access, approve, deny, check-in/out)
-
-Workflow commands are accessed via: pam workflow <subcommand>
-"""
-
 __all__ = ['PAMWorkflowCommand', 'check_workflow_access', 'check_workflow_and_prompt_2fa']
 
-from .workflow_commands import PAMWorkflowCommand, check_workflow_access, check_workflow_and_prompt_2fa
-
+from .registry import PAMWorkflowCommand
+from .mfa import check_workflow_access, check_workflow_and_prompt_2fa
